rename isprinklr.context to isprinklr.paths which is more descriptive
