+++ isprinklr/scheduler.py
@@ -9,6 +9,8 @@ import time
 import pandas as pd
 import requests
 
+from isprinklr.paths import logs_path, config_path
+
 API_URL = "http://localhost:8080/api/"
 
 day_abbr = {
@@ -26,7 +28,7 @@ day_abbr = {
 def read_schedule():
     """Reads the sprinkler schedule from a .csv file and returns a list of dictionaries"""
     try: 
-        df = pd.read_csv("../data/schedule.csv", usecols=["zone", "day", "duration"])
+        df = pd.read_csv(data_path + "/schedule.csv", usecols=["zone", "day", "duration"])
         return df, False
     except Exception as e:
         logging.error(f"Unable to open schedule.csv: {e}")
@@ -140,7 +142,7 @@ def main():
 
 
 if __name__ == "__main__":
-    logging.basicConfig(handlers=[RotatingFileHandler('../logs/scheduler.log', maxBytes=1024*1024, backupCount=1, mode='a')],
+    logging.basicConfig(handlers=[RotatingFileHandler(logs_path + '/scheduler.log', maxBytes=1024*1024, backupCount=1, mode='a')],
                     format='%(asctime)s %(levelname)s: %(message)s',
                     level=logging.DEBUG) 
     logger = logging.getLogger()
